llm-node/requirements.txt
fastapi
uvicorn
pydantic
pycryptodome
pynacl
